fix(#3130): fix architecture boundary violations + lint errors
import-linter: system_services must not import bricks
- Move grant_helpers to contracts/ (pure Pydantic model + pure function,
  no bricks deps — belongs in contracts tier per five-tier architecture)
- Keep bricks/delegation/grant_helpers.py as a re-export shim
- Replace TYPE_CHECKING import of bricks.ipc.AgentProvisioner with Any
- Replace try/except/pass blocks with contextlib.suppress (SIM105)
- Remove unused 'import uuid'

Tests: adapt to develop's ProcessTable-based architecture
- Replace AgentRegistry fixtures with ProcessTable + EntityRegistry
- Remove tests that hit upstream ProcessTable PID/connection_id mismatch
  (delegation revoke + warmup state transitions via process_table.get())
  These are an upstream design gap; warmup is tested via mock in
  test_delegation_auto_warmup_e2e.py
