+++ tests/unit/bricks/delegation/test_grant_helpers.py
@@ -2,7 +2,7 @@
 
 import pytest
 
-from nexus.bricks.delegation.grant_helpers import (
+from nexus.contracts.grant_helpers import (
     MAX_REGISTRATION_GRANTS,
     GrantInput,
     grants_to_rebac_tuples,
